Updates to work with rpm version 4.12.0.1-11 from RedHat
As of that version, rpm does not read the passphrase from stdin and pass it to gpg via fd 3. Instead the passphrase is read from the gpg-agent. If the gpg-agent does not have the passphrase, it can be configured via option "allow-loopback-pinentry", to get it from gpg instead of a separate pinentry program. If gpg is invoked with "--pinentry-mode loopback", then the passphrase can be provided by jenkins via expect/send as before.

Updated the expect script to exit on eof, which will occur if gpg-agent does not require a passphrase (likely because it is cached). Use "log_user 0" to try to prevent the passphrase from ending up in the jenkins build output on an error.